feat: Add authentication state management and splash screen for new users only (#11)
- Add auth storage utilities (AsyncStorage) for session management
- Update app entry to check auth state and route accordingly
- Logged-in users go directly to home screen (skip splash/auth)
- New users see splash screen during signup flow only
- Existing users skip splash and welcome screens
- Add SplashScreen component for new user onboarding
- Update AuthFlow to show splash only for new users
- Save authentication state after successful login/signup
- Install @react-native-async-storage/async-storage
