+++ Lumio/components/auth/AuthFlow.tsx
@@ -4,9 +4,11 @@ import { AmbientBackground } from './ambient/AmbientBackground';
 import { EntryScreen } from './EntryScreen';
 import { PasswordScreen } from './PasswordScreen';
 import { WelcomeScreen } from './WelcomeScreen';
+import { SplashScreen } from './SplashScreen';
 import { AuthColors, AuthTypography } from '@/constants/auth-theme';
+import { AuthStorage } from '@/utils/auth-storage';
 
-type AuthStep = 'entry' | 'password' | 'welcome' | 'complete';
+type AuthStep = 'splash' | 'entry' | 'password' | 'welcome' | 'complete';
 type UserType = 'new' | 'existing' | null;
 
 interface AuthFlowProps {
@@ -52,6 +54,7 @@ export function AuthFlow({ onComplete }: AuthFlowProps) {
   const [email, setEmail] = useState('');
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
+  const [showSplashForNewUser, setShowSplashForNewUser] = useState(false);
 
   const handleEmailSubmit = async (submittedEmail: string) => {
     setEmail(submittedEmail);
@@ -78,14 +81,17 @@ export function AuthFlow({ onComplete }: AuthFlowProps) {
       setUserType(type);
 
       if (type === 'existing') {
-        // Existing user - go directly to app
+        // Existing user - save auth and go directly to app
+        await AuthStorage.saveAuth('oauth_token_' + provider, 'oauth@user.com', false);
         setStep('complete');
         setTimeout(() => {
           onComplete();
         }, 1500);
       } else {
-        // New user - show welcome screen
-        setStep('welcome');
+        // New user - save auth and show splash screen first
+        await AuthStorage.saveAuth('oauth_token_' + provider, 'oauth@user.com', true);
+        setShowSplashForNewUser(true);
+        setStep('splash');
       }
     } catch (err) {
       setError('Authentication failed. Please try again.');
@@ -102,13 +108,19 @@ export function AuthFlow({ onComplete }: AuthFlowProps) {
       if (userType === 'new') {
         const result = await createAccount(email, password);
         if (result.success) {
-          setStep('welcome');
+          // Save auth state for new user
+          await AuthStorage.saveAuth('token_' + Date.now(), email, true);
+          // Show splash screen for new user
+          setShowSplashForNewUser(true);
+          setStep('splash');
         } else {
           setError('Account creation failed. Please try again.');
         }
       } else {
         const result = await authenticatePassword(email, password);
         if (result.success) {
+          // Save auth state for existing user
+          await AuthStorage.saveAuth('token_' + Date.now(), email, false);
           setStep('complete');
           setTimeout(() => {
             onComplete();
@@ -139,6 +151,18 @@ export function AuthFlow({ onComplete }: AuthFlowProps) {
     }, 3000);
   };
 
+  const handleSplashComplete = () => {
+    // After splash, show welcome screen for new users
+    if (userType === 'new') {
+      setStep('welcome');
+    } else {
+      setStep('complete');
+      setTimeout(() => {
+        onComplete();
+      }, 1500);
+    }
+  };
+
   const handleWelcomeComplete = () => {
     setStep('complete');
     setTimeout(() => {
@@ -148,36 +172,42 @@ export function AuthFlow({ onComplete }: AuthFlowProps) {
 
   return (
     <View style={styles.container}>
-      <AmbientBackground />
+      {step === 'splash' && showSplashForNewUser ? (
+        <SplashScreen onComplete={handleSplashComplete} />
+      ) : (
+        <>
+          <AmbientBackground />
 
-      {step === 'entry' && (
-        <EntryScreen
-          onEmailSubmit={handleEmailSubmit}
-          onOAuthPress={handleOAuthPress}
-          loading={loading}
-        />
-      )}
+          {step === 'entry' && (
+            <EntryScreen
+              onEmailSubmit={handleEmailSubmit}
+              onOAuthPress={handleOAuthPress}
+              loading={loading}
+            />
+          )}
 
-      {step === 'password' && userType && (
-        <PasswordScreen
-          email={email}
-          userType={userType}
-          onPasswordSubmit={handlePasswordSubmit}
-          onBack={handleBack}
-          onForgotPassword={handleForgotPassword}
-          loading={loading}
-        />
-      )}
+          {step === 'password' && userType && (
+            <PasswordScreen
+              email={email}
+              userType={userType}
+              onPasswordSubmit={handlePasswordSubmit}
+              onBack={handleBack}
+              onForgotPassword={handleForgotPassword}
+              loading={loading}
+            />
+          )}
 
-      {step === 'welcome' && (
-        <WelcomeScreen onEnter={handleWelcomeComplete} />
-      )}
+          {step === 'welcome' && (
+            <WelcomeScreen onEnter={handleWelcomeComplete} />
+          )}
 
-      {step === 'complete' && (
-        <View style={styles.completeContainer}>
-          <Text style={styles.completeText}>Entering your space…</Text>
-          <ActivityIndicator size="large" color={AuthColors.primary} style={{ marginTop: 24 }} />
-        </View>
+          {step === 'complete' && (
+            <View style={styles.completeContainer}>
+              <Text style={styles.completeText}>Entering your space…</Text>
+              <ActivityIndicator size="large" color={AuthColors.primary} style={{ marginTop: 24 }} />
+            </View>
+          )}
+        </>
       )}
 
       {error && (
